Make cfile compile on OSX with s/uint64_t/size_t and fat lock in bloomfile

+++ src/kudu/cfile/block_pointer.h
@@ -45,7 +45,7 @@ class BlockPointer {
   string ToString() const {
     char tmp[100];
     snprintf(tmp, sizeof(tmp), "offset=%ld size=%d",
-             offset_, size_);
+             (long) offset_, size_);
     return string(tmp);
   }
